refactor: chain requests lifecycle for fragment matching with then handler
