+++ packages/web-fragments/src/gateway/index.ts
@@ -1,2 +1,2 @@
 export { FragmentGateway } from './fragment-gateway';
-export type { FragmentConfig } from './utils/types';
+export type { FragmentConfig } from './middleware/types';
