+++ packages/web-fragments/src/gateway/middleware/types.d.ts
