+++ packages/web-fragments/vite.config.ts
@@ -17,7 +17,6 @@ export default defineConfig({
 				gateway: new URL('src/gateway/index.ts', import.meta.url).pathname,
 				'gateway/node': new URL('src/gateway/middleware/node.ts', import.meta.url).pathname,
 				'gateway/web': new URL('src/gateway/middleware/web.ts', import.meta.url).pathname,
-				utils: new URL('src/gateway/utils/common-utils.ts', import.meta.url).pathname,
 				elements: new URL('src/elements/index.ts', import.meta.url).pathname,
 			},
 			formats: ['es'],
